add addDbTreeHeader function, fix issue when autocomplete button disappears on database change
