+++ pgmanage/app/static/assets/js/pgmanage_frontend/src/tab_functions/outer_connection_tab.js
@@ -34,7 +34,8 @@ import {
   refreshHeights,
   resizeConnectionHorizontal,
   resizeTreeVertical,
-  toggleTreeContainer
+  toggleTreeContainer,
+  toggleConnectionAutocomplete
 } from "../workspace";
 import { beforeCloseTab } from "../create_tab_functions";
 import { connectionsStore } from "../stores/connections.js";
@@ -456,4 +457,26 @@ function refreshOuterConnectionHeights() {
   }
 }
 
-export { v_createConnTabFunction, refreshOuterConnectionHeights }
+function addDbTreeHeader(header_div, tab_id, database_name) {
+  let autocomplete_btn_id = `autocomplete_toggler_${tab_id}`
+  let autocomplete_switch_status =
+    settingsStore.enableAutocomplete !== false
+      ? " checked "
+      : "";
+
+  header_div.innerHTML = `<i class="fas fa-server mr-1"></i>selected DB:
+      <b>${database_name}</b>
+      <div class="omnidb__switch omnidb__switch--sm float-right" data-toggle="tooltip" data-placement="bottom" data-html="true" title="" data-original-title="<h5>Toggle autocomplete.</h5><div>Switch OFF <b>disables the autocomplete</b> on the inner tabs for this connection.</div>">
+        <input type="checkbox" ${autocomplete_switch_status} id="${autocomplete_btn_id}" class="omnidb__switch--input">
+        <label for="${autocomplete_btn_id}" class="omnidb__switch--label">
+              <span>
+                  <i class="fas fa-spell-check"></i>
+              </span>
+          </label>
+  </div>`;
+  
+  let autocomplete_btn = document.getElementById(`${autocomplete_btn_id}`)
+  autocomplete_btn.onchange = function() { toggleConnectionAutocomplete(autocomplete_btn_id) }
+}
+
+export { v_createConnTabFunction, refreshOuterConnectionHeights, addDbTreeHeader }
