+++ pgmanage/app/static/assets/js/pgmanage_frontend/src/tree_context_functions/tree_mysql.js
@@ -27,7 +27,6 @@ SOFTWARE.
 */
 import { createApp } from "vue";
 import TreeMysql from "../components/TreeMysql.vue";
-import { toggleConnectionAutocomplete } from "../workspace";
 import { tabSQLTemplate } from "./tree_postgresql";
 import { createMessageModal } from "../notification_control";
 import { refreshMonitoring } from "../tab_functions/inner_monitoring_tab";
@@ -35,8 +34,8 @@ import { showPasswordPrompt } from "../passwords";
 import { execAjax } from "../ajax_control";
 import axios from "axios";
 import { showToast } from "../notification_control";
-import { settingsStore } from "../stores/settings";
 import { emitter } from "../emitter";
+import { addDbTreeHeader } from "../tab_functions/outer_connection_tab";
 
 /// <summary>
 /// Retrieving tree.
@@ -353,27 +352,8 @@ function getTreeMysql(div) {
   // save tree referece in the tab, it will be later used to destroy tree instance on tab close
   v_connTabControl.selectedTab.tree = app;
 
-  let autocomplete_switch_status =
-    settingsStore.enableAutocomplete !== false
-      ? " checked "
-      : "";
-  let autocomplete_btn_id = `autocomplete_toggler_${v_connTabControl.selectedTab.tag.tab_id}`;
-
-  v_connTabControl.selectedTab.tag.divDetails.innerHTML = `<i class="fas fa-server mr-1"></i>selected DB:
-        <b>${v_connTabControl.selectedTab.tag.selectedDatabase}</b>
-        <div class="omnidb__switch omnidb__switch--sm float-right" data-toggle="tooltip" data-placement="bottom" data-html="true" title="" data-original-title="<h5>Toggle autocomplete.</h5><div>Switch OFF <b>disables the autocomplete</b> on the inner tabs for this connection.</div>">
-    	    <input type="checkbox" ${autocomplete_switch_status} id="${autocomplete_btn_id}" class="omnidb__switch--input" onchange="toggleConnectionAutocomplete(\'autocomplete_toggler_${v_connTabControl.selectedTab.tag.tab_id}\')">
-    	    <label for="${autocomplete_btn_id}" class="omnidb__switch--label">
-                <span>
-                    <i class="fas fa-spell-check"></i>
-                </span>
-            </label>
-		</div>`;
-
-  let autocomplete_btn = document.getElementById(`${autocomplete_btn_id}`);
-  autocomplete_btn.onchange = function () {
-    toggleConnectionAutocomplete(autocomplete_btn_id);
-  };
+  let tab_tag = v_connTabControl.selectedTab.tag
+  addDbTreeHeader(tab_tag.divDetails, tab_tag.tab_id, tab_tag.selectedDatabase)
 }
 
 /// <summary>
